+++ src/c/methods/callbacks/MeCallbacks.cpp
@@ -69,7 +69,7 @@ void modioOnGetUserEvents(u32 call_number, u32 response_code, json response_json
     {
       if(modio::hasKey(response_json, "data"))
       {
-        events_array_size = response_json["data"].size();
+        events_array_size = (u32)response_json["data"].size();
         events_array = new ModioEvent[events_array_size];
 
         for(u32 i=0; i<events_array_size; i++)
